kleinere Layout-Fixes

+++ src/de/osmembrane/view/panels/PipelinePanel.java
@@ -86,9 +86,9 @@ public class PipelinePanel extends JPanel implements Observer {
 	/**
 	 * The constants applicable for the layeredPane.
 	 */
-	private static final Integer FUNCTION_LAYER = new Integer(1);
+	private static final Integer FUNCTION_LAYER = new Integer(3);
 	private static final Integer CONNECTOR_LAYER = new Integer(2);
-	private static final Integer LINK_LAYER = new Integer(3);
+	private static final Integer LINK_LAYER = new Integer(1);
 
 	/**
 	 * Saves the point in object coordinates when a drag and drop action occurs
